requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-core==1.30.0
azure-data-tables==12.5.0
azure-functions==1.18.0
certifi==2024.2.2
charset-normalizer==3.3.2
courlan==1.0.0
dateparser==1.2.0
distro==1.9.0
h11==0.14.0
htmldate==1.7.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
isodate==0.6.1
jusText==3.0.0
langcodes==3.3.0
lxml==5.1.0
multidict==6.0.5
openai==1.12.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-telegram-bot==20.8
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tld==0.13
tqdm==4.66.2
trafilatura==1.7.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
